Refactor Rcpp function getPileups for speed
Replace vector iteration with a std::map container for faster searching. Also, reduce the number of rows in overlaps to only distinct rows, but mutate in a column to track the number of duplications that existed for that row. Utilize this in the new getPileupsMap() function to cut down on loop iterations.