Update plot_gtf to call diff sorting methods based on features present
Once the initial input gtf data frame is split into strand specific data frames, there is a possibility that one or more of the resulting data frames will no longer have the features that the original sorting methods expected to be present. Now each strand specific data frame will call the sorting method it needs for the features it has.