Fix issue where arc plots were created with old data
helix.txt is used to create arc plots, and MiSiPi is checking for the presence of this file to know if it should generate this plot. The file was not being deleted in between iterations, however, leading to a situation where an arc plot would occasionally be generated for a locus that should not have one. This commit addresses that.